requirements.txt
pandas==2.2.3
numpy==2.2.5
scipy==1.15.3
permetrics==2.0.0
streamlit==1.45.1
plotly==6.1.0
lmstudio==1.3.0
st-pages==1.0.1
streamlit-option-menu==0.4.0
openai==1.86.0
python-dotenv==1.1.0